+++ app/src/main/java/com/example/taskmaster/MainApplication.kt
@@ -1,0 +1,7 @@
+package com.example.taskmaster
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class MainApplication : Application()
